feat: adding lazy load to 'Game' class. Building with errors (can't find 'calculateWinner' func at 'Board')
