+++ .gitignore
@@ -11,6 +11,9 @@
 # production
 /build
 
+# versioning
+/package-lock.json
+
 # misc
 .DS_Store
 .env.local
