+++ src/components/Square.js
@@ -1,0 +1,7 @@
+export function Square(props) {
+    return (
+      <button className="square" onClick={props.onClick}>
+        {props.value}
+      </button>
+    );
+  }
